+++ pkg/domain/repository/articlesRepository.go
@@ -1,0 +1,4 @@
+package repository
+
+type Articles interface {
+}
