+++ server.go
@@ -2,16 +2,15 @@ package main
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/sawadeeeen/sawaya-api/internal/apps/article-api/createArticle"
-	"github.com/sawadeeeen/sawaya-api/internal/apps/article-api/getArticles"
+	"github.com/sawadeeeen/sawaya-api/internal/apps/articlesApi"
 )
 
 func main() {
 	router := gin.Default()
 	v1 := router.Group("/v1")
 	{
-		v1.GET("/", getArticles.GetArticles)
-		v1.GET("/create", createArticle.CreateArticle)
+		v1.GET("/", articlesApi.GetArticles)
+		v1.POST("/create", articlesApi.CreateArticle)
 		// v1.GET("/:id", getArticleById)
 		// v1.POST("/:id", editArticle)
 		// v1.POST("/", deleteArticle)
